Prepare tests environment for Erlang zones
Up to this point the test environment was assuming
zones were being created from Java image.

That is no longer tha case, so I had to perform
various changes in the script and in the code to resolve
the issue. Now any image meeting certain requirements
can be used as zones.

As K8s have changed since June 2021, I had to change
our yaml-s in order for the job's volume claim to work.

Another important change is that now tester's job
creates and delete zones for itself.
Thus the only thing that a person needs to do in order
to run the test is to set variables in test-config.conf
file and apply kustomization.yaml file.

The last change is as follows: I've deleted redundant
EXECUTOR zone (Java zone that was acting like a client)
because job can run Java code itself, as it contains
required app.jar.